change styling to a prop
